Integrate SelfDisposingMessage into Message. This has almost no implication for the overhead benchmark.
